+++ E_paper.h
@@ -7,8 +7,8 @@
 //  if you are not sure which version, please test each one,
 //  if it is successful then it belongs to the model of the file name
 // The 2.66 screen has other SPI pins, and more pixels. Partof the pixels are now unused.
-#include <GxDEPG0213BN/GxDEPG0213BN.h>  //Huidige schermen, 212*104
-//#include <GxGDEH0213B73/GxGDEH0213B73.h>  //Eerste schermen, 212*104
+//#include <GxDEPG0213BN/GxDEPG0213BN.h>  //Huidige schermen, 212*104
+#include <GxGDEH0213B73/GxGDEH0213B73.h>  //Eerste schermen, 212*104
 //#include <GxGDEM0213B74/GxGDEM0213B74.h>  // 2.13", 212*104 
 //#include <GxDEPG0266BN/GxDEPG0266BN.h>  //2.66 screen for rollz...
 //#include <GxGDEW0213M21/GxGDEW0213M21.h>  //nog een andere versie, resolutie NIET OK vanwege resolutie !!!
@@ -98,6 +98,7 @@ extern bool SoftAP_connection;
 extern bool GPS_Signal_OK;
 extern bool Shut_down_Save_session;
 extern float RTC_voltage_bat,alfa_window;
+extern float RTC_minimum_voltage_bat;
 extern  double delta_heading,ref_heading;
 extern int bootCount,run_count,stat_count,GPIO12_screen;
 extern int ftpStatus;