+++ gpx.h
@@ -28,9 +28,12 @@ SOFTWARE.
 //gpx 1.0 is used here ! 
 //https://logiqx.github.io/gps-wizard/gpx/
 //Always 1Hz points
+#include "Ublox.h"
 #define GPX_HEADER 0
 #define GPX_FRAME 1
 #define GPX_END 2
+extern const char SW_version[16];
+//extern UBXMessage ubxMessage; 
 void log_GPX(int part,File file){
 char bufferTx[512]; 
 int i,y;